requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.1.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-materialize==1.0.1a2
django-materializecss-form==1.1.17
djangorestframework==3.11.0
docopt==0.6.2
Flask==1.1.1
geographiclib==1.50
geopy==1.20.0
gmplot==1.2.0
googletrans==2.4.0
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Js2Py==0.66
key==0.4
MarkupSafe==1.1.1
numpy==1.17.2
packaging==19.2
pandas==0.25.1
Pillow==6.0.0
pipwin==0.4.5
psycopg2-binary==2.8.2
pyjsparser==2.7.1
pyparsing==2.4.2
PyPrind==2.11.2
python-dateutil==2.8.0
python-googlegeocoder==0.4.0
pytz==2018.9
requests==2.21.0
robobrowser==0.5.3
six==1.12.0
soupsieve==1.7.3
SpeechRecognition==3.8.1
tzlocal==2.0.0
urllib3==1.24.1
virtualenv==16.6.0
Werkzeug==0.16.0
whitenoise==5.0.1
wikipedia==1.4.0
